requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
earthengine-api
black
